removing tutorial info
